+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Fixed
 - List type with empty list default causes failure `PyLaia#48
   <https://github.com/jpuigcerver/PyLaia/issues/48>`__.
 - Pure dataclass instance default being considered as a subclass type.
+- Discard ``init_args`` after ``class_path`` change causes error `#205
+  <https://github.com/omni-us/jsonargparse/issues/205>`__.
 
 
 v4.18.0 (2022-11-29)